Classes de pesquisa atualizadas

+++ Classes/ValorItemComparator.java
@@ -1,10 +1,13 @@
+import java.util.Comparator;
+
+import itens.Item;
 
 public class ValorItemComparator implements Comparator<Item> {
 	
 	@Override
 	public int compare(Item item, Item outroItem) {
-		Integer item1 = item.getValor();
-		Integer item2 = outroItem.getValor);
+		Double item1 = item.getValor();
+		Double item2 = outroItem.getValor();
 		return item1.compareTo(item2);
 	}
